[ansible_plugins] Don't override comment state
This patch fixes an issue where the filter plugin would override the
state of a value with 'comment' state to a 'present' state if another
entry with the same value is defined elsewhere. Now, the plugin will
check if a given value has 'comment' state and the new value also has
'comment' state, and will not change that state in that case.