[debops.proc_hidepid] Check /proc ownership
In addition to checking current POSIX capabilities, the role will check
if the '/proc' directory is owned by 'root' before trying to remount it.
This improves DebOps support for unprivileged LXC containers, where the
remount is skipped since the '/proc' directory is most likely owned by
'nobody' instead of 'root'.