[debops.owncloud] Fix evaluation of `gpg --verify` exit code
Before the exit code of `gpg` was never directly evaluated but the one of
`sed` which is the last command in the pipe.

Fortunately, ypid does not see as way how this could have been used to
circumvent signature checking by the role because the output still
needs to contain the full OpenGPG fingerprint. Anyway, this should not
have happend.