[debops.nodejs] Change default NPM version
The 'latest' branch in NPM's GitHub repository seems to be an equivalent
of a stable branch.

This change is prompted by the issues with Etherpad installation which
stops with an error message from NPM:

    npm WARN deprecated socks@1.1.10: If using 2.x branch, please
    upgrade to at least 2.1.6 to avoid a serious bug with socket data
    flow and an import issue introduced in 2.1.0

I'm not sure if installing latest stable NPM version directly
system-wide is a good idea. This is hopefully a short-lived solution, in
the future I want to implement NPM installation per-user in case of
GitHub based installation. The 'debops.nodejs' role would require an
user account to install NPM in, and otherwise skip the NPM installation
entirely.