[debops.owncloud] Fix permission issue of /srv/www which prevented ACME
The issue was that the Nextcloud data directory was created by the `env`
role which run before one of the webserver roles run. The data directory
is set restrictive and the Ansible `file` module created the parent
directories with the same permissions which is not desirable.

I am not sure anymore why I created the data directory in the `env` role
but I deployed Nextcloud 13 against a fresh Debian Stretch without
issues with the change of this commit.

Closes: #275