Update 'debops.gitlab' test in GitLab CI
The UNIX groups created by the 'debops.auth' role are now managed by the
'debops.system_groups' role which is a soft dependency of the
'debops.gitlab' role (included in the playbook).

The test environment should now include the 'acl' APT package which
provides the filesystem ACL support. There shouldn't be a need to
specifically disable this feature for testing.