Improve Python 3 support in the common playbook
The roles that are used in the 'common.yml' playbook will have a new set
of Ansible local facts compatible with Python 3 interpreter. The new
facts are activated when Ansible uses Python 3.x interpeter, otherwise
the facts which use Python 2.7 will be active. This allows Python 3-only
Debian environment in certain conditions.