[debops.ifupdown] Add 'debops.kmod' dependency
Some of the network interface types (bridges, bonding, VLANs) require
additional kernel modules to be loaded at boot time or during
configuration. This can be handled by the operating system on demand,
however this results in warnings about missing modules and can cause
issues at boot time.

This patch adds the 'debops.kmod' role as a dependency of the
'debops.ifupdown' role. The 'debops.ifupdown' role will check what
network interfaces are to be configured on a host, and will dynamically
generate configuration for the 'debops.kmod' role which will then load
the missing kernel modules and mark them to be loaded at boot time.