[debops.postgresql] Escape '*' in regexp match
The 'debops.postgresql' role uses the 'lineinfile' Ansible module to
manage the contents of the '~/.pgpass' files. Some of the fields in the
file can contain '*' characters, however in the regexp matching the '*'
character is special and needs to be escaped to be treated literally.

This fixes the issue with the 'debops.postgresql' role idempotency.