[debops.postgresql] Synchronize password lookups
The 'postgresql__password_hostname' variable in the 'debops.postgresql'
role is used to select correct path in the 'secret/' directory to
retrieve the password. It was used by the task that created the
PostgreSQL roles in the database, but in the task that saved the
password in the '~/.pgpass' file. This could result in wrong password
being used for authenticated connections to the database located on
localhost, depending on the host inventory name.

This patch synchronizes the 'secret/' path used for PostgreSQL role
creation and path to the password stored in '~/.pgpass' file, which
should fix the issue.