Implement upstream version checks using 'uscan'
This patch adds a way to check the upstream version of the third-party
software managed by DebOps which is not included in Debian Archive, or
distributed via additional APT repositories. This will be used to
monitor changes and new releases of upstream software to update the
version used for installation by DebOps roles.

The check can be done by executing the command:

    make versions

in the root of the DebOps monorepo. The 'uscan' tool from the Debian
'devscripts' APT package is required.