[debops.sshd] Don't block SSH access if standalone
The 'debops.sshd' role checks the Ansible local facts of the
'debops.system_groups' role to get the list of UNIX groups which are
allowed to login and configure them in the 'AllowGroups' parameter.

This change ensures that if the above facts are not set, ie.
'debops.sshd' role is used as standalone, the 'AllowGroups' will not be
set - otherwise this would result in locked down system access where
only 'root' account can login, which might not be enabled.