[debops.mariadb] Flatten dependent configuration
The configuration specified via dependent role variables is usually
passed as a "list of lists of items", therefore some tasks that used
"with_together" lookup didn't work properly with more than one item.
This patch ensures that all lists are flattened to the same level, which
should fix the issue.