[debops.postgresql_server] Don't change template1
The 'template1' PostgreSQL database is used as a template for newly
created databases. The 'debops.postgresql_server' role contained a set
of tasks that modified the 'template1' database to allow better support
for multi-tenancy in the PostgreSQL service, based on documentation from

  https://wiki.postgresql.org/wiki/Shared_Database_Hosting

Unfortunately it seems that the changes made by the role affect
installation of packaged database schemas which are not designed with
these modifications in mind. In some cases this results in broken usage
of 'dbconfig' functionality in Debian. Due to that, the 'template1'
modifications are removed from the base 'debops.postgresql_server' role.
They might be re-introduced in DebOps in a separate Ansible role that
will support PostgreSQL multi-tenancy in the future.