[debops.nginx] Update 'try_files' to support lists
The 'item.try_files' directive used in the 'debops.nginx' server
configuration can be specified either as a string or as a list. The
templates that generate the configuration will now check which version
is used and correctly format the result. This avoids an issue where
a string is treated as a list and formatted incorrectly with a space
between each letter.