[debops.ferm] Handle IPv4/IPv6 addresses in 'dmz'
The 'ferm' language does not support iterating over lists or converting
lists to strings, this caused an issue where a list of IP addresses is
concatenated with a string representing a port when the 'dmz' rule type
is used. The result was an error:

    variable 'PRIVATE_IP' must be a string, but it is an array

This patch changes the template for the 'dmz' rule so that the IPv4 and
IPv6 addresses are separated on Ansible level instead of directly in
ferm. Next, only 1 public and private IP address is selected for the
'dmz' rule so that the string concatenation with a port can be performed
correctly. This should solve the issue with the 'dmz' rule type.