[debops.ferm] Add 'snat_ip' parameter to rules
Based on the changes made by Gaudenz Steinlin in commit 29863982.

This patch changes how the SNAT firewall configuration is implemented in
the 'debops.ferm' role, 'dmz' rule type. With this set, connections to
the backend host will have the source IP set to this address instead of
the originating IP. The original rule in the POSTROUTING chain is not
needed as this case is already automatically covered by the connection
tracking done by the DNAT rule in the PREROUTING chain.