[debops.ansible_plugins] Change how defaults work
Previously if 'parse_kv_items' filter invocation specified default
parameters to set using the 'defaults' option, and the configuration
being filtered contained an alternative set of parameters, the defaults
provided as the filter option were discarded. Now, they will be added to
the list of parameters and will be present in addition to the normal
options. The defaults can be overriden or removed as usual by use of the
'state' parameter.