[debops.kmod] Fix usage in Ansible --check mode
The role would result in an error when it was run on a host for the
first time in the Ansible '--check' mode due to missing local facts. The
tasks that load/unload the kernel modules will now only be activated
when the 'ansible_local.kmod' local facts are present.