[debops.bootstrap] Allow removal of 127.0.1.1 IP
The 127.0.1.1 IP address is rutinely used in '/etc/hosts' of standalone
systems to denote the hostname and/or domain of the "current system".
This is useful in absence of a given host entry in the DNS database.

In distributed systems this might cause problems when hosts pass their
own IP addresses between themselves - suddenly a given IP address from
a remote host refers to "current system" and the communication between
hosts is broken.

The 'debops.bootstrap' role can now remove the 127.0.1.1 entries from
the '/etc/hosts' files added by provisioning services to avoid this
issue. This allows for correct usage when the host entry is also defined
in the DNS database to allow domain and hostname resolution.