[debops.core] Update the 'resolver.fact' script
The 'resolver.fact' script is converted into a Python script to make the
data gathering and JSON output more sane.

If the system resolver is configured to use local nameserver on the
'127.0.0.1' address, the 'resolver.fact' will expose the upstream
nameservers managed by the 'resolvconf' APT package in a separate list.
These can be used in applications that configure DNS nameservers on
different "hosts", for example Docker containers.