[debops.lxc] Redesign how containers are created
Originally, new LXC containers created by the 'debops.lxc' role had to
always be restarted to use the new MAC address generated by the
'lxc-hwaddr-static' script. This causes the DHCP server to see two
consecutive MAC addresses on the network and may cause confusion or
other unnecessary issues.

This patch modifies the container creation behavior slightly: if the
'item.state' parameter is not explicitly specified, the role will create
the new LXC container in the "stopped" state, generate static MAC
addresses and start the container. This does not require an explicit
container restart for it to use the new MAC addresses and is faster.

If the 'item.state' is explicitly specified to 'started', the container
will be started as before, and then restarted to use the new MAC
addresses generated by the LXC pre-start hook.