[debops.lxc] Remove 127.0.1.1 host entry
This patch ensures that the LXC containers created using Ansible tasks
in the 'debops.lxc' role don't have the '127.0.1.1' entries in their
'/etc/hosts' files. This is done to ensure that the LXC containers
created using the Ansible tasks as well as the 'lxc-new-unprivileged'
script are equivalent.

The '127.0.1.1' entry is used to resolve the local hostname, but it
might interfere in the multi-host environment with DHCP/DNS services
present. Because the LXC environment maintained by DebOps includes
a local DNS/DHCP services, these entries are not needed in '/etc/hosts'.