[debops.netbase] Ensure POSIX capabilities support
The 'debops.netbase' role depends on checking for 'cap_sys_admin'
POSIX capability to change the hostname. Ansible requires the
'libcap2-bin' package to be present on the remote host before exposing
list of POSIX capabilities in host facts.

This patch ensures that the 'libcap2-bin' package is installed before
being used by the role.