[debops.lxc] Manage LXC containers via systemd
Due to issues with LXC containers not being shut down properly via the
'lxc-stop' command, the LXC environment managed by 'debops.lxc' will
switch to using the systemd 'lxc@.service' instances to start/stop LXC
containers. This allows for the containers to be stopped "from the
inside" via the 'lxc-attach' command and ensure that the containerized
services are correctly shut down.