[ci] Add a way to force IPv4/IPv6 use by APT
During testing of DebOps roles, some of package installs result in APT
trying to connect to IPv6 address of a mirror instead of IPv4 and
failing due to not available IPv6 connectivity - the issue is the DNS
resolution that returns IPv6 addresses with higher priority than IPv4.

This patch adds an option for the 'Vagrantfile' configuration to create
the VMs/containers that force APT to use IPv4 or IPv6 network. This can
be enabled in the test environment to avoid issues like the one above
during DebOps tests.