[debops.owncloud] Avoid privilege escalation
The 'unarchive' task in the 'debops.owncloud' role was executed as
an unprivileged user, however the specified 'owner' and 'group'
parameters forced the module to request privilege escalation to root.
This doesn't work because the 'www-data' UNIX account under which the
task is executed is not authorized for privilege escalation. The task
exited with the "Timeout (12s) waiting for privilege escalation prompt"
error.

To fix this, the 'unarchive' task is executed as privileged, and the
'unarchive' module will use its own mechanism to change the owner and
group of extracted files.