[debops.lxc] Include parent domain in 'lxc-net'
If the DNS domain used in the LXC environment has at least 3 levels and
its parent domain has at least 2 levels, the parent domain will be
included in the list of searchable domains inside of the LXC containers.
This should allow for easy resolution of hosts in the parent domain from
inside of the internal LXC containers connected to the 'lxcbr0' bridge.