Import 'dnsmasq' playbook in 'lxc' playbook
The 'lxc.yml' playbook will import the 'dnsmasq.yml' playbook. In case
that the 'debops.dnsmasq' role is enabled on the LXC host, this allows
to apply the needed changes in the 'dnsmasq' configuration to allow
resolution of the LXC container addresses on the LXC host.

If the 'debops.dnsmasq' role is not enabled on the LXC host, this
shouldn't change anything.