Remove ansible dependency in the 'debops' script.
All interaction with ansible are done with python's subprocess,
which finds ansible in the env. PATH.
The version-check in 'debops' script however, previously relied on
the ansible python-library, which isn't guaranteed to be the same
version as ansible found in the PATH.

This change will make the version-check with the printout from
'ansible-playbook --version' command instead.

The other nice benefit from this is that debops doesn't have a hard
dependency on ansible, making it much lighter.