[debops.netbase] Implement automatic domain config
The 'debops.netbase' role will check the FQDN and IP address of a given
host from that host's perspective to determine if its current domain and
IP address are defined locally via the '/etc/hosts' database or via
a DNS service. This will be used to update the '/etc/hosts' database.

The existing '127.0.1.1' entry in the local host database will be
removed, and if DNS service is not available, a local FQDN and hostname
will be defined in '/etc/hosts' which will point to the host's default
IPv4 and IPv6 addresses. The domain will be generated automatically
based on the 'ansible_host' value, or if it's an IP address, on the
'inventory_hostname' value.