[debops.netbase] Fix undefined 'ansible_ssh_host'
When the 'debops.netbase' role is executed against local host, the
'ansible_host' and 'ansible_ssh_host' variables may be undefined. In
that case, short-circuit the condition with "0" which will be
interpreted as an IP address and should avoid selecting the wrong value
as the host FQDN and domain name.