[debops.netbase] Keep the domain defined locally
If the 'debops.netbase' role determines that the host's FQDN defined in
the '/etc/hosts' file points to a different IP address than '127.0.1.1',
it will assume that the DNS domain is defined locally and will not try
to remove the existing entries. This should preserve the hostname and
FQDN configuration in environments without a working DNS service.