[debops.dnsmasq] Change the rebind domain treshold
The 'debops.dnsmasq' role will now check if the base domain used by the
host has 4 or more levels instead of 3 or more to allow rebinding IP
addresses on its parent domain. This is done to avoid allowing rebinds
on 1st and 2nd level gTLDs which are common in country top-level
domains.