[debops.ifupdown] Stop, disable systemd-networkd
The 'systemd-networkd' service will try to configure the network
interfaces independently of the ifupdown scripts. If the user enabled
the 'debops.ifupdown' role on a host we can assume that the ifupdown
scripts should manage the network interfaces instead. Therefore the role
will stop and disable the 'systemd-networkd' service to avoid its
interference.