[Vagrant] Redesign hostname & DNS configuration
The previous network configuration used in Vagrant tests relied on the
external DNS and DHCP services present in the test environment. This
resulted in issues with the Vagrant test machines and required
additional external setup of the test environment.

This patch changes the Vagrant configuration to use an internal
'.vagrant.test' DNS domain for the test virtual machines. In addition,
the Vagrantfile provisioning scripts will fix the issue with 'vagrant'
configuring the test machine hostname in '/etc/hosts' file to point to
the '127.0.0.1' address - instead, the host FQDN and hostname will point
to its default IP address which is what DebOps playbooks and roles
expect.

The test machine hostname will contain a randomized component to allow
multiple test machines to exist at the same time and not clash in the
DNS and hostname resolution.

The configuration of an additional public network for the test machines
is also dropped to avoid issues with misconfigured external test
environment.