[ci] Improve test inventory variables
This patch changes the dynamically generated test Ansible inventory in
the 'jane' script as well as the Vagrantfile to use the real host FQDN
instead of 'localhost' as the name of the host managed by Ansible. The
test inventory will also add the IP address of the host from which the
Vagrant SSH connects to the list of Ansible Controllers to ensure that
Vagrant parent host is not cut off by the firewall configuration.