[ci] Reorder PostgreSQL and Redis Server playbooks
It seems that installation of Redis Server changes the value of the
'/proc/sys/kernel/shmmax' kernel parameter which prompts the PostgreSQL
role to reconfigure the database and results in the GitLab test being
idempotent. Installing the Redis Server before PostgreSQL Server solves
the problem, but the reason will have to be investigated to avoid this
requirement in the future.