[debops.gitlab] Don't hard-depend on 'debops.pki'
The 'debops.pki' role does not have to be included as a dependency of
the 'debops.gitlab' role, because it is exepcted to be applied on the
hosts already.

Alternatively, it would be expected to be added to the role playbook as
a "soft" dependency instead of the role 'meta/main.yml' file as a "hard"
dependency. Any communication between roles can be faciliated via
Ansible local facts.