[debops.docker_registry] Redesign token support
The support for token-based authentication in 'debops.docker_registry'
role was focused too much on support for GitLab Container Registry
service, therefore it was updated to be service-agnostic. The GitLab
support is enabled by default when the GitLab installation is detected
via Ansible local facts, but other "providers" can be used as well, with
easy switch between them.