[debops.docker_registry] Allow access to storage
The access to the Docker Registry storage directory is required by the
GitLab backup script to create backups of stored container images. Since
it was either allowing access to all UNIX accounts on the host, or
adding the 'git' UNIX account to the 'docker-registry' UNIX group,
I chose the former method as the default one.