[debops.gitlab] Don't generate Pages configuration
If the 'GitLab Pages' functionality is not enabled by the
'debops.gitlab' role, the 'debops.nginx' role will not generate a custom
configuration file with the server options for it. This is done to avoid
generating nginx configuration file with an empty 'server_name' which
could be mistakenly enabled in the webserver and break the existing
setup.