[debops.docker_registry] Clean Redis cache on GC
When Docker Registry performs garbage collection, it does not remove
existing data cached in Redis. This might cause issues with the Registry
when users try to push the same images a second time. Because of that,
when any garbage collection happens, the script will remove all Redis
keys related to Registry to avoid such issues.