[debops.nginx] Add global list of hostname domains
The 'nginx__hostname_domains' variable can be used to specify the list
of DNS domains which should be used as suffixes for short hostnames. The
first found domain specified there which matches the currently selected
domain will be used as the base domain. This can be overridden by
specifying the desired base domain in the 'hostname_domain' parameter.