[debops.apt_proxy] Don't validate via external URL
The role doesn't have to validate the Python script against an external
URL not controlled by the user each time the script is deployed - that
should be done in the DebOps CI pipeline during tests, not in the
production environment. Additionally if we were to do this, the URL
needs to be configurable by the user via a default variable.