[debops.slapd] Correctly mark single-valued attrs
LDAP specification defaults to multi-value entries, and single-value
entries need to be specifically defined in the schema. The convention in
the 'debops.slapd' configuration should be the same - any attributes
defined without a special 'state' parameter will be assumed as
multi-value, but single-value attributes should be defined with the
'state' parameter set as 'exact'.

This will carry on to the configuration specified in the Ansible
inventory, therefore users will not have to explicitly set the state on
the entries defined in the role defaults that are single-value.