[debops.slapd] Switch to shared superuser password
In a multi-master replication mode, the 'cn=config' and main LDAP
database superupser passwords will be replicated between the nodes. To
avoide mismatche and update loops, the role will use a shared superuser
passwords for these databases by default. The passwords are stored in
plaintext on the Ansible Controller in the 'secret/slapd/' subdirectory.