[debops.slapd] Change password storage scheme
The 'password' Ansible lookup plugin doesn't work very well with spacess
in the filenames. Because LDAP DNs can contain spaces and other
characters that might cause issues in the future, the passwords stored
in the 'secret/' directory will be saved in files with names based on
the UUIDs generated from the LDAP Distinguished Names.