[debops.cryptsetup] Fixes present state flag behavior
The state flag of the plaintext block devices controls if the block
devices must be opened or not and if the filesystem on it must be
mounted.

In case the filesystem is not managed by the role and the state flag is
set to `present', the plaintext block device should not be closed as it may be
used by other roles like debops.lvm.

Then, this commit adds an additional condition inside the teardown procedure
at the end of the role in order to implemente this use case.

In a further patch, the lvm-based use case should be better covered with
explicit configuration other than simply disable filesystem management.