[debops.slapd] LDAP is a "public" service
By default the 'debops.slapd' role will configure OpenLDAP as a "public"
service, meaning that anybody from the Internet can connect to it
through the firewall and TCP Wrappers. Default variables can be used to
control this behaviour and limit connections to specific set of IP
addresses or CIDR subnets.