[ci] Add mock LDAP servers in Vagrant tests
The 'vagrant.test' domain inside of the Vagrant test VMs is not a real
DNS domain. The 'debops.ldap' role uses 'dig' Ansible lookup plugin to
get the address of the LDAP servers, which doesn't work very well during
tests. For now, replace that lookup with a static list of LDAP servers.