[debops.users] Drop the support for prefixes
The support for prefixing the UNIX groups and accounts with '_'
character when in the LDAP environment has been removed from the role,
to make it easier to manage application accounts via the Ansible
inventory. The feature will be implemented in a separate Ansible role
meant to be used to manage local administrator and user accounts.