[debops.core] Install 'dirmngr' by default
The 'dirmngr' package is required by GnuPG to perform network
operations, like downloading and importing GPG keys from keyservers.
it's not required by default in Debian, and might not be present on
minimal installations, but it might be needed by other DebOps roles to
function properly, therefore the 'debops.core' role will ensure that it
is present by default.

Ref: https://bugs.debian.org/845720
Ref: https://blog.sleeplessbeastie.eu/2017/11/02/how-to-fix-missing-dirmngr/