[debops.sudo] Allow dpkg to remove 'sudo' package
The 'SUDO_FORCE_REMOVE=yes' variable tells dpkg(1) that it can safely
remove the 'sudo' package, which can happen when we try to switch to the
'sudo-ldap' package.

Normally 'dpkg' checks if the 'root' account has a password and will
stop the upgrade process because access to the 'root' account could be
lost when the password is not set; but we assume that users can still
access root via SSH public keys and loss of access is unlikely to happen
in this case.