[debops.elasticsearch] Divert packaged sysctl conf
The 'elasticsearch' .deb package comes with packaged sysctl
configuration in the /usr/lib/sysctl.d/ directory. The configuration
file modifies the virtual memory parameter of the Linux kernel, which is
usually read-only in a container environment. This breaks the use of the
'sysctl' command to update kernel parameters for a given container.

The 'debops.elasticsearch' role will divert the file that comes with the
.deb package, and the 'debops.syctl' role will be used to apply the
relevant virtual memory configuration. This might result in the
Elasticsearch service not starting correctly in unprepared container,
but the playbook run will be correct.

(cherry picked from commit daa914d1f4551b3631d7364d942589575d53ca77)