[debops.keyring] Use 'root' instead of omit
For some reason, Ansible from the 'stretch-backports' Debian repository
fails with the error "TypeError: 'NoneType' object is not iterable" when
the 'become_user' value should be marked as omitted. Setting the 'root'
account explicitly makes the error go away and is an equivalent action.

This happens on Ansible 2.7.5 from Debian, Ansible from upstream seems
to be unaffected.