[debops.mariadb_server] Don't set root password
This patch removes the 'root' password configuration code from the
'debops.mariadb_server' role.

Currently with Ansible 2.8, setting a 'root' MariaDB account password
with the 'mysql_user' module breaks access to the database. But setting
a custom password is not really needed anymore due to the 'unix_socket'
authentication plugin configured for the UNIX 'root' account.