[debops.system_groups] Track the connection type
The 'ansible_connection' variable is not correctly resolved inside of
Jinja templates - it turns into a 'ansible.utils.sentinel.Sentinel'
Python object instead. Because of that, if we want to use it directly or
indirectly inside of Jinja templates, it needs to be resolved
beforehand, for example as a runtime fact.