[debops.python] Disable Pyv2 without Ansible facts
This is a revert of commit 30d03e1d122a8d4800e6e4c676bba04d064e568b

The previous commit introduced an unintended effect that the role did
not correctly recognized use of Python 3 as the Ansible interpreter in
the 'raw' mode - Python3-only environment was not enforced. Because of
that, the 'debops.system_users' role requested execution of a Python
script on the Ansible Controller using Python 2.7, which didn't work on
Ansible Controller with Python3-only environment.

This change might require intentional activation of the Python 2.7
environment in the inventory via the 'ansible_python_interpreter'
variable, but with move to Python 3.x this shouldn't be needed much
longer.