[debops.lxc] Add base DNS domain to resolvconf
In addition to the 'lxc.<domain.org>' DNS domain, the 'debops.lxc' role
will ensure that the '<domain.org>' DNS domain is also included in the
'/etc/resolv.conf' file via the 'resolvconf' package. This ensures that
the "virtual" '<container>.lxc' DNS domain is also resolvable from the
LXC host.