[debops.docker_server] Don't depend on ferm facts
The 'debops.docker_server' wanted to install 'ferment' if the 'ferm'
firewall is installed by the 'debops.ferm' role; however on an initial
run without the 'ferm' firewall configured the condition was False, and
on subsequent runs it changed to True when the Ansible local facts
became available.

There's currently no good solution to this issue (detecting ferm
beforehand), so for now the 'debops.docker_server' role will install the
'ferment' package and corresponding 'ferm' rules by default, if the
upstream Docker packages are disabled (relevant to older versions of
Docker).