[debops.python] Better tracking of environment
On initial run of the 'debops.python' role on Debian Buster with Python
2.7 installed, the role would assume that the Python 2.7 is not
installed and disable support for it. On subsequent runs, the facts
would tell the role that Python 2.7 is present and support for it would
be enabled, changing the state of the host by installing missing Python
2.7 packages.

This patch ensures that the role knows the actual state of Python 2.7
and Python 3.x environment on the host on the initial run, and the
enabled/disabled state is correctly defined for subsequent runs.