[debops.keyring] Check if UNIX account exists
If the GPG keys are supposed to be managed on an unprivileged UNIX
account, check if that account actually exists before trying to manage
them. This might not be the case when upstream installation requires
a separate UNIX account but installation using OS package does not
because the APT package will create that account automatically.