[debops.dnsmasq] Add 'lxc' host record if needed
When the LXC support is enabled in the 'debops.dnsmaq' role and there's
no DNS domain configured (standalone host), the 'lxc' subdomain and
corresponding 'lxcbr0' network interface cannot be resolved from the LXC
host itself due to lack of 'search' entries in the '/etc/resolv.conf'
file. To fix that, the 'debops.dnsmasq' role will create a static host
record that points to the 'lxcbr0' network interface.