[debops.libvirtd] Reorder roles for idempotency
In the 'debops.libvirtd' playbook, the 'debops.nsswitch' role needs to
be applied before the 'debops.ferm' role, so that the 'debops.ferm' role
knows aboout the Avahi service installed prior, without the use of the
'debops.avahi' role. This fixes an issue with idempotent playbook
execution, where on the second run the 'debops.ferm' role "learns" about
Avahi being present via the 'nsswitch' local Ansible fact and configures
access for Avahi service, thus breaking the idempotency.

This is a quirk in the DebOps CI pipeline, but it might have application
in production environments as well.