Remove 'debops.root_account' from bootstrap plays
The 'debops.root_account' role, among other things, modifies the 'root'
account password to one stored in the 'secret/' directory on the Ansible
Controller. If that happens when Ansible manages the host via the 'root'
account with a password (SSH keys are not present), Ansible cannot
connect to the host anymore and its execution fails.

The 'bootstrap.yml' and 'bootstrap-ldap.yml' playbooks are commonly used
via the 'root' account to prepare the host for Ansible management. If
the host was provisioned without setting SSH keys for the 'root'
account, due to behaviour described above Ansible will fail its
execution when the host is provisioned for the first time.

Because of the above, the 'debops.root_account' role will be removed
from the bootstrap playbooks. Its presence in the 'common.yml' playbook
is fine, because at this point we can assume that the system
administrator will use the admin account prepared with the bootstrap
playbook to manage the host, and changing the 'root' password during
execution shouldn't produce any ill effects.

(cherry picked from commit 59168b9923fff871ff4390f8f0f07038779b7db2)