Fix DMZ ferm rule type with dport set
The filter rule in the FORWARD chain is traversed after the DNAT rule
in the PREROUTING chain. Thus the filter rule has to match on the
dport value instead of the port value.

This issue was already fixed in commit 7e247611ccb56476a86766c2707d198da1346c32
for the old dmz rule template.