[debops.golang] Reload facts on app installations
This patch fixes an issue where configuring the 'debops.golang' on
a host and then installing a Go application on the same host in
a separate playbook did not ensure that the Go application binary
location was available via Ansible local facts. Now, facts will be
updated when a Go application is installed to get its location into the
'debops.golang' facts.