[dpkg_divert] Rename the diverted files by default
When a diversion is requested, usually the intention is to move an
existing file (or a potential future file) out of the way to replace it
with a modified version. This patch changes the module behavior to
rename the existing files by default, so that in most common cases this
does not have to be specifically requested. Renaming can still be
disabled by setting the 'rename' parameter to False.