[dpkg_divert] Rename the diverted files by default
When a diversion is requested, usually the intention is to move an
existing file (or a potential future file) out of the way to replace it
with a modified version. This patch changes the module behavior to
rename the existing files by default, so that in most common cases this
does not have to be specifically requested. Renaming can still be
disabled by setting the 'rename' parameter to False.

(cherry picked from commit 691d9219288de28201ae0a77dd0cb0e1a821e83c)