[dpkg_divert] Implement deletion of changed files
This patch adds support for the 'delete' parameter in the 'dpkg_divert'
Ansible module. Adding this parameter in the task will tell the module
to remove the changed file before reverting the original, thus saving us
from having to remove the changed file in a previous task.

The 'delete' parameter name was chosen because 'dpkg-divert' command
already has a '--remove' argument which is used for a different purpose.