[debops.cron] Redesign randomization algorithm
The patch changes the 'debops.cron' /etc/crontab randomization algorithm
to be a bit more random and stores the seeds used to randomize execution
times on the host via facts. The changes aren't ideal, the values still
sometimes repeat on a given host more than I would like but it should do
for now.