[debops.postldap] Domain lookup include subdomains
This patch adds support for matching subdomains of the host DNS domain
in LDAP search filters. This should ensure that the mail senders cannot
use a known subdomain of the host's primary DNS domain to send
unauthorized mail messages.

This only works with one domain at a time, if multiple DNS domains are
used in the LDAP directory, the relevant search queries need to be
updated to account for all of the possible domain levels.