[debops.slapd] Block access to directory service
This patch changes the default access policy to the OpenLDAP service
from "allow from anywhere" to "block by default", with an option to
allow access to the service using the 'slapd__*_allow' variables.

This change should limit exposure of the LDAP directory and the data
contained in it only to the trusted subnets instead of the entire
Internet. Access from outside of the trusted networks should be granted
using a VPN service.